website-backup/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
authlib>=1.4.0
httpx>=0.28.0
itsdangerous>=2.2.0
starlette>=0.45.0
mangum>=0.19.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
